pdf parser

+++ city/fetch.py
@@ -146,24 +146,17 @@ class ProjectFetcher():
                                 attachment.attachment.save(title, pdf_file, True)
                                 attachment.save()
                                 
-                                # firstpage = PdfFileReader(pdf_file).getPage(0)
-                                # print firstpage
                                 try:
                                     img= Image(blob=pdf_file)
                                     img = Image(img.sequence[0])
-                                    print img.size
                                     img.format='png'
-                                    # img.resize(300,420)
                                     img_url = settings.MEDIA_ROOT + "project_attachments/images/"+ img_name
                                     img.save(filename=img_url)
-                                    print img
-
-                                    # result = urllib2.urlopen('file://' +img_url).read()
-                                    # print result
                                     attachment.image = "project_attachments/images/"+ img_name
                                     attachment.save()
                                 except wand.exceptions.DelegateError:
-                                    print 'DelegateError'
+                                    print 'DelegateError', link
+                           
                             os.remove('tmp_pdf')
                 
                 except lxml.etree.XMLSyntaxError:
@@ -307,29 +300,45 @@ class ProjectFestgestelltFetcher():
             if link:
                 #scrape pdfs from project page
                 project_page = requests.get(link)
-                tree = html.fromstring(project_page.content)
-                # print tree
-                # pdfs = tree.xpath('//a[@class="pdf"]/@href')
-                pdfs = tree.xpath('//a[@class="nscout pdf"]')
-                # print link
-                # print 'pdfs: ', pdfs
-                for pdf in pdfs:
-                    pdf_link = 'http://www.hamburg.de'+ pdf.xpath('@href')[0]
-                    title = pdf.xpath('text()')[0].lstrip().split('(')[0][:-1]
+                try:
+                    tree = html.fromstring(project_page.content)
+                    pdfs = tree.xpath('//a[@class="nscout pdf"]')
 
-                    pdf_name = pdf_link.split('/')[-1]
+                    for pdf in pdfs:
+                        pdf_link = 'http://www.hamburg.de'+ pdf.xpath('@href')[0]
+                        
+                        #check if attachment already exists in database
+                        if not ProjectAttachment.objects.filter(source=pdf_link).exists():
+                            title = pdf.xpath('text()')[0].lstrip().split('(')[0][:-1]
 
-                    raw_pdf = urllib2.urlopen(pdf_link)
-                    attachment = ProjectAttachment(name=pdf_name, project=project)
+                            pdf_name = pdf_link.split('/')[-1]
+                            img_name = pdf_name.split('.pdf')[0] + '.png'
 
-                    with open('tmp_pdf', 'wb') as f:
-                        f.write(raw_pdf.read())
+                            raw_pdf = urllib2.urlopen(pdf_link)
+                            attachment = ProjectAttachment(name=pdf_name, project=project, source=pdf_link)
 
-                    with open('tmp_pdf', 'r') as f:
-                        pdf_file = File(f)
-                        attachment.attachment.save(pdf_name, pdf_file, True)
-                        attachment.save()   
-                    os.remove('tmp_pdf')
+                            with open('tmp_pdf', 'wb') as f:
+                                f.write(raw_pdf.read())
+
+                            with open('tmp_pdf', 'r') as f:
+                                pdf_file = File(f, 'r')
+                                attachment.attachment.save(title, pdf_file, True)
+                                attachment.save()
+                                
+                                try:
+                                    img= Image(blob=pdf_file)
+                                    img = Image(img.sequence[0])
+                                    img.format='png'
+                                    img_url = settings.MEDIA_ROOT + "project_attachments/images/"+ img_name
+                                    img.save(filename=img_url)
+                                    attachment.image = "project_attachments/images/"+ img_name
+                                    attachment.save()
+                                except wand.exceptions.DelegateError:
+                                    print 'DelegateError', pdf_link
+                            os.remove('tmp_pdf')
+                
+                except lxml.etree.XMLSyntaxError:
+                    print 'XMLSyntaxError for: ', link
 
 
             date = feature['properties'].get('feststellung', '')
